Sprockets 2 compatibility
