+++ src/prototype/ajax/periodical_updater.js
