+++ src/prototype/ajax/responders.js
