+++ src/prototype/ajax/updater.js
