+++ src/prototype/deprecated.js
