+++ src/prototype/lang.js
@@ -1,3 +1,18 @@
+//= compat
+//= require "lang/class"
+//= require "lang/object"
+//= require "lang/function"
+//= require "lang/date"
+//= require "lang/regexp"
+//= require "lang/periodical_executer"
+//= require "lang/string"
+//= require "lang/template"
+//= require "lang/enumerable"
+//= require "lang/array"
+//= require "lang/hash"
+//= require "lang/number"
+//= require "lang/range"
+
 /**
  * == Language ==
  * Additions to JavaScript's "standard library" and extensions to
@@ -16,23 +31,23 @@ var Abstract = { };
  *
  *  Accepts an arbitrary number of functions and returns the result of the
  *  first one that doesn't throw an error.
- *  
+ *
  *  **This method is deprecated.**
  *
  *  <h5>More information</h5>
  *
- *  [[Try.these]] provides a simple idiom for trying out blocks of code in 
- *  sequence. Such a sequence of attempts usually represents a downgrading 
+ *  [[Try.these]] provides a simple idiom for trying out blocks of code in
+ *  sequence. Such a sequence of attempts usually represents a downgrading
  *  approach to obtaining a given feature.
  *
- *  In this example from Prototype's [[Ajax section]] internals, we want to get an 
- *  `XMLHttpRequest` object. Internet Explorer 6 and earlier, however, does not 
- *  provide it as a vanilla JavaScript object, and will throw an error if we 
- *  attempt a simple instantiation. Also, over time, its proprietary way 
+ *  In this example from Prototype's [[Ajax section]] internals, we want to get an
+ *  `XMLHttpRequest` object. Internet Explorer 6 and earlier, however, does not
+ *  provide it as a vanilla JavaScript object, and will throw an error if we
+ *  attempt a simple instantiation. Also, over time, its proprietary way
  *  evolved, changing COM interface names.
  *
- *  [[Try.these]] will try several ways in sequence, from the best (and, 
- *  theoretically, most widespread) one to the oldest and rarest way, returning 
+ *  [[Try.these]] will try several ways in sequence, from the best (and,
+ *  theoretically, most widespread) one to the oldest and rarest way, returning
  *  the result of the first successful function.
  *
  *  If none of the blocks succeeded, [[Try.these]] will return `undefined`, which
@@ -64,17 +79,3 @@ var Try = {
     return returnValue;
   }
 };
-
-//= require "lang/class"
-//= require "lang/object"
-//= require "lang/function"
-//= require "lang/date"
-//= require "lang/regexp"
-//= require "lang/periodical_executer"
-//= require "lang/string"
-//= require "lang/template"
-//= require "lang/enumerable"
-//= require "lang/array"
-//= require "lang/hash"
-//= require "lang/number"
-//= require "lang/range"
