+++ src/prototype/lang/array.js
