+++ src/prototype/lang/class.js
