+++ src/prototype/lang/enumerable.js
