+++ src/prototype/lang/function.js
