+++ src/prototype/lang/hash.js
