+++ src/prototype/lang/number.js
