+++ src/prototype/lang/object.js
