+++ src/prototype/lang/periodical_executer.js
