+++ src/prototype/lang/range.js
