+++ src/prototype/lang/regexp.js
