+++ src/prototype/lang/string.js
