+++ src/prototype/lang/template.js
